An s4 class for Control object for stomach compilation.

An s4 class for Control object for stomach compilation.

Value

object of class STOMcontrol

Slots

name

character. Name of control object.

stomach_format

expression (R code) for file name for CSV file defining stomach contents data. Default: expression(system.file("stomach_format.csv", package = "FishStomachs")) For more details on exchange format see the vignette: vignette("Database_format", package = "FishStomachs")

stomach_dir

characters. Directory name with stomach data on exchange format.

dataSets

File name or file names for data on exchange format.

years

Year or years of data to be used.

quarters

Quarters of the year ( or similar time split) to be used.

predators

Predator name or names.

strata_area_exp

expression for defining strata area.

strata_sub_area_exp

expression for defining sub_area.

strata_time_exp

expression for defining temporal strata.

strata_year_back

expression for getting year information out of diet data, e.g. expression(as.numeric(substr(stratum_time,1,4))).

strata_quarter_back

expression for getting quarter information out of diet data.

sel_preys

Selected prey names for diet calculation.

min_prey_length

Minimum prey length (or included in other food) integer.

calc_strata

Specification of strata

weighting_factor

Weighting factor for weighted mean

year_quarter_strata

vector.

weighting_by_nstom

logical.

combine_starta_from_relative_stomach_content

logical.

model_options

list. Options for producing input values to multi species models as SMS.

The list must as a minimum include the names :
 "delete_small" logical, delete input with relative stomach contents weight lower than "min_value".
                if FALSE values lower than "min_value" are increased to "min_value".
 "min_value", minimum relative stomach contents weight (see above)
 "insert_mid_value" logical, insert dummy value ("mid_value") where observations for a prey  within a size range are missing
 "mid_value" value used for missing mid-values.
 "insert_tails", logical for insertion of dummy values("tail_value") for the prey size class which are lower and higher than the observed size range for a prey
 "tail_value",  value used for missing tail-values.
 

max_l

integer. Number used for the maximum length, e.g. 9998L

mis_l

integer. Number used for missing length, e.g. 9999L

mis_digest

integer. Number used for missing digestion stage e.g. 9L

other

character. Name for the "Other food", which is prey items not considered as individual prey species for further analysis, e.g. "other".

mis_size_class

integer. Number for size class where the size is unknown, e.g. 0L

ALK_options

list, options used for creating Age Length Keys

do_test_output

logical.

do_test_output_file

logical.

tst_file

character.

detailed_tst_output

logical.

detailed_tst_file

character.

detailed_tst_criteria

list.

boots_id

list of expression for defining sampling unit (haul) from STOMobs variables.

boots_strata

list expression for defining strata (haul).

Examples

if (FALSE) control<-new("STOMcontrol",
      name='Baltic cod',
      stomach_dir=file.path(stom_data_dir,'baltic'),
      dataSets= c("balt_stom_new_formatNO_new_var.csv"))