Statistics from bootstrap replicates.

Statistics from bootstrap replicates.

bootsMean(
  b,
  pointEst,
  by_prey_size = FALSE,
  n_rep = NA,
  do_Diri = FALSE,
  minPreyProportion = 0,
  Diri_min = 0.001
)

Arguments

b

Bootstrap diet data set. List of elements of class STOMdiet.

pointEst

Diet data estimated without bootstrapping (point estimates of prey weights)

by_prey_size

Logical for calculating mean and variance by prey size. FALSE provides statistics by prey species.

n_rep

Selected replicates used. If n_rep is a number, 1:n_rep are used, If n_rep is a vector the provided replica numbers are used in the statistics. All replicates are uesed if n_rep is NA (default).

do_Diri

Logical for performing statistics assuming the replicates are Dirichlet distributed.

minPreyProportion

Lower level for average prey proportion to be kept as a named prey. If lower, the prey weight is allocated to "other" prey.

Diri_min

Value to be used in estimating the Dirichelt parameter in case of a zero proportion.

Value

Data set with mean and variance of diets and fit to Dirichlet distribution. Output includes

Variable	Contents

nboots	Number of bootstrap replicates.

year	Year.

quarter	Quarter of the year.

pred_name	Predator name.

pred_size	Predator size.

prey	Prey, prey name and size, if used.

prey_name	Prey name.

prey_size	Prey size.

n_prey_sp	Number of "preys" observed.

param	Estimated parameters assuming a Dirichlet distribution of bootstrap replicates.

phi	Estimated precision parameter (sum of param from the Dirichlet distribution).

mu	Estimated mean value from the Dirichlet distribution.

p_value	p-value for Log-likelihood ratio test for the estimated Dirichlet

mean vector describe the bootstrap data well.

mean_w	Mean value of prey weights (proportions).

sd_w	Standard deviation of prey weights (proportions).

prey_w	Point estimate of prey weight (proportions).

strata	Strata used for bootstrap.

n_stom	Number of stomachs available before bootstrap ( from the pointEst data).

n_sample	Number of samples available before bootstrap.